+++ +euphonic/Contents.m
@@ -1,0 +1,13 @@
+% +euphonic
+%
+% A matlab wrapper for the Python package Euphonic to calculate phonon spectra from DFT
+% calculations, including functions to integrate with the Horace INS data analysis program.
+%
+% Files
+%   CoherentCrystal        - Matlab wrapper around a Python class to calculate the
+%                            coherent neutron intensity S(q,w) for single crystals
+%   ForceConstants         - Matlab class around the ForceConstants Python class
+%   mp_grid                - Returns a Monkhorst-Pack grid of specified size
+%   ureg                   - Matlab wrapper around the ureg units registry
+%   install_python_modules - A script to install (using `pip`) the required Python modules
+