+++ +euphonic/private/required_modules.m
@@ -1,0 +1,3 @@
+function out = required_modules()
+    out = {'euphonic', 'TO_BE_DETERMINED'};
+end