+++ get_hor_eu_interface_version.m
@@ -1,3 +1,0 @@
-function ver = get_hor_eu_interface_version()
-	ver = char(py.get_hor_eu_interface_version.get_version());
-end