+++ test/EuphonicTest.m
@@ -14,29 +14,31 @@ classdef EuphonicTest < EuphonicTestSuper
     
     methods (TestMethodSetup, ParameterCombination='sequential')
         function set_euphonic_sf_args(testCase, use_c, n_threads, chunk)
-            qpts = testCase.qpts;
             opts = testCase.opts;
             pars = testCase.pars;
-            phonon_kwargs = {'phonon_kwargs', ...
-                                {'asr', 'reciprocal', ...
-                                    'use_c', use_c, ...
-                                    'n_threads', n_threads}};
+            phonon_kwargs = {'asr', 'reciprocal', ...
+                             'use_c', use_c, ...
+                             'n_threads', n_threads};
             opts = [opts phonon_kwargs];
             if ~ismissing(chunk)
                 opts = [opts {'chunk', chunk}];
             end
+            opts = [opts {'scattering_lengths', testCase.scattering_lengths}];
+            opts = [opts {'temperature', pars(1)}];
             
-            euphonic_sf_args = {qpts(:, 1), qpts(:, 2), qpts(:, 3), ...
-                                pars, testCase.scattering_lengths, opts};
-            testCase.euphonic_sf_args = euphonic_sf_args;
+            testCase.euphonic_sf_args = opts;
         end
     end
 
     methods(Test, ParameterCombination='sequential', TestTags={'integration'})
         function runIntegrationTests(testCase)
-            [w, sf] = euphonic_sf(testCase.euphonic_sf_args{:});
-            w_mat = cell2mat(w);
-            sf_mat = cell2mat(sf);
+            coherentsqw = euphonic.CoherentCrystal(testCase.force_constants, ...
+                                                   testCase.euphonic_sf_args{:});
+            qpts = testCase.qpts;
+            [w, sf] = coherentsqw.horace_disp(qpts(:, 1), qpts(:, 2), qpts(:, 3), ...
+                                              testCase.pars(2));
+            w_mat = transpose(cell2mat(w'));
+            sf_mat = transpose(cell2mat(sf'));
 
             fname = get_expected_output_filename(testCase.material_name, ...
                                                  testCase.pars, testCase.opts);
@@ -44,8 +46,12 @@ classdef EuphonicTest < EuphonicTestSuper
             expected_w_mat = cell2mat(expected_w);
             expected_sf_mat = cell2mat(expected_sf);
 
-            testCase.verifyTrue( ...
-                all(ismembertol(w_mat, expected_w_mat, 1e-5), 'all'));
+            import matlab.unittest.constraints.IsEqualTo
+            import matlab.unittest.constraints.AbsoluteTolerance
+            import matlab.unittest.constraints.RelativeTolerance
+            bounds = AbsoluteTolerance(5e-4) | RelativeTolerance(0.01);
+            testCase.verifyThat(w_mat, ...
+                IsEqualTo(expected_w_mat, 'within', bounds));
             % Ignore acoustic structure factors by setting to zero - their
             % values can be unstable at small frequencies
             sf_mat(:, 1:3) = 0;
@@ -59,8 +65,9 @@ classdef EuphonicTest < EuphonicTestSuper
             % Need to sum over degenerate modes to compare structure factors
             sf_summed = sum_degenerate_modes(w_mat, sf_mat);
             expected_sf_summed = sum_degenerate_modes(w_mat, expected_sf_mat);
-            testCase.verifyTrue( ...
-                all(ismembertol(sf_summed, expected_sf_summed, 1e-2), 'all'));
+            bounds = AbsoluteTolerance(0.01) | RelativeTolerance(0.01);
+            testCase.verifyThat(sf_summed, ...
+                IsEqualTo(expected_sf_summed, 'within', bounds));
         end
     end
 end