+++ CHANGELOG.rst
@@ -1,3 +1,27 @@
+`Unreleased <https://github.com/pace-neutrons/horace-euphonic-interface/compare/v0.1.0...HEAD>`_
+----------
+
+There has been a major refactor, which means changes to how
+Horace-Euphonic-Interface is installed. There are also major changes
+to how Euphonic is used, the API has been updated to make it more
+object-oriented.
+
+- Installation changes:
+ - Horace-euphonic-interface is now distributed as a Matlab toolbox (``.mltbx``)
+   which is available in the `Matlab File Exchange <https://www.mathworks.com/matlabcentral/fileexchange/>`_ as an Add-On
+
+- Usage changes:
+ - ``euphonic_sf`` has been removed
+ - ``euphonic_on`` has been removed
+ - Force constants are now a separate object (``ForceConstants``) rather than
+   passing these arguments to ``euphonic_sf``
+ - The model parameters are set in a ``CoherentCrystal`` model object, rather than
+   passing these parameters to ``euphonic_sf``
+ - The function handle to be passed to ``disp2sqw_eval`` is ``CoherentCrystal.horace_disp`` rather than ``euphonic_sf``
+ - The ``dw_grid`` argument has been renamed to ``debye_waller_grid``
+
+For more detailed help see the `documentation <https://horace-euphonic-interface.readthedocs.io/en/latest/>`_
+
 `v0.1.0 <https://github.com/pace-neutrons/horace-euphonic-interface/compare/81607231b...v0.1.0>`_
 ------